Running hypothesis.
Performing the cleaning of outliers and after that divided's the data set into two different sets according to admit column and plotting QQplot and histogram for the values gpa and gre. To get the overview of these two values for admitted and rejected students. Then perform t-test, Anova test and chi square test to check the dependency of columns.